abi: Enable support for 64-bit ABI
This change updates ABI in Hubris to support 64-bit. This involves
two pieces:
1. Host
2. Runtime

For runtime, it is straightforward since `usize` can be used to
determine how many bytes it takes to reference any memory location
and the registers.

For host, `usize` does not provide this information. Thus, we have to
instead rely on inputs from user to determine what ABI should be used
for building any host tools e.g. xtask.

Thus, sys/abi is updated to define AbiSize as one of the following:
1. usize --> Required for runtime by kernel and user tasks
2. u32   --> Required for host tools that use 32-bit ABI
3. u64   --> Required for host tools that use 64-bit ABI

In order to enable the above definitions, following two feature flags
are added that can be selected by the user when building host tools:
`host_use_abi32` and `host_use_abi64`. This change updates sys/abi and
all dependent host tools (xtask), kern and user tasks (including tests)
to update the data types used in ABI calls.

+++ sys/kern/src/syscalls.rs
@@ -570,8 +570,8 @@ fn borrow_lease(
         // Now, proceed only if both neither the length nor address computation
         // wrap.
         if let (Some(off_len), Some(off_addr)) = (
-            lease.length.checked_sub(offset),
-            lease.base_address.checked_add(offset),
+            lease.length.checked_sub(offset as usize),
+            lease.base_address.checked_add(offset as usize),
         ) {
             lease.base_address = off_addr;
             lease.length = off_len;
